Continued styling across password reset/forgot and requesting another email pages
Moved color variables to base/variables file.
Moved styles that apply to whole document to the custom.scss file.
Changed button color to be brown-grey. This is changed in the base/buttons file.
Aligned the buttons centrally using the containing div and text align. This is more reliable than the columns as the buttons are different widths.
Increased height of navigation bar as it was slighlty smaller than the image.